Written all unit tests for application layer

+++ GoodBadHabitsTracker.Application/Commands/Habits/DeleteAllProgress/DeleteAllProgressCommandHandler.cs
@@ -1,26 +1,29 @@
-﻿using GoodBadHabitsTracker.Application.Commands.Habits.DeleteAll;
-using GoodBadHabitsTracker.Application.Exceptions;
+﻿using GoodBadHabitsTracker.Application.Exceptions;
 using GoodBadHabitsTracker.Core.Interfaces;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using LanguageExt.Common;
+using System.Net;
 
 namespace GoodBadHabitsTracker.Application.Commands.Habits.DeleteAllProgress
 {
     internal sealed class DeleteAllProgressCommandHandler(
         IHabitsRepository habitsRepository,
-        IUserAccessor userAccessor) : IRequestHandler<DeleteAllProgressCommand, bool>
+        IUserAccessor userAccessor) : IRequestHandler<DeleteAllProgressCommand, Result<bool>>
     {
-        public async Task<bool> Handle(DeleteAllProgressCommand command, CancellationToken cancellationToken)
+        public async Task<Result<bool>> Handle(DeleteAllProgressCommand command, CancellationToken cancellationToken)
         {
-            var user = await userAccessor.GetCurrentUser()
-                ?? throw new AppException(System.Net.HttpStatusCode.Unauthorized, "User not found");
+            var user = await userAccessor.GetCurrentUser();
+            if (user == null)
+                return new Result<bool>(new AppException(HttpStatusCode.BadRequest, "User Not Found"));
+            
             var userId = user.Id;
 
-            return await habitsRepository.DeleteAllProgressAsync(userId, cancellationToken);
+
+            await habitsRepository.DeleteAllProgressAsync(userId, cancellationToken);
+            /*var postInProgressTodayTask = habitsRepository.PostInProgressTodayAsync(allHabits, cancellationToken);*/
+
+
+            return new Result<bool>(true);
 
 
         }
